+++ frontend/src/App/component/sonar/BugsPage.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
-import ProjectAvatar from './ProjectAvatar'
+import ProjectAvatar from './../ProjectAvatar';
 import Axios from 'axios'
 import { CircularProgress, Backdrop } from '@material-ui/core'
 import { connect } from 'react-redux'
-import DrawingBoard from './DrawingBoard'
+import DrawingBoard from './../DrawingBoard'
 import moment from 'moment'
 
 const useStyles = makeStyles((theme) => ({
@@ -21,16 +21,14 @@ const useStyles = makeStyles((theme) => ({
   },
 }))
 
-function DuplicationsPage(prop) {
+function BugsPage(prop) {
   const classes = useStyles()
-  const [duplicationList, setDuplicationList] = useState([])
+  const [bugList, setBugList] = useState([])
   const [currentProject, setCurrentProject] = useState(undefined)
-  const [duplicationUrl, setDuplicationUrl] = useState("")
-  const [dataForDuplicationChart, setDataForDuplicationChart] = useState({ labels: [], data: { duplication: [] } })
+  const [bugUrl, setBugUrl] = useState("")
+  const [dataForBugChart, setDataForBugChart] = useState({ labels: [], data: { bug: [] } })
 
   const projectId = localStorage.getItem("projectId")
-  const jwtToken = localStorage.getItem("jwtToken")
-
   const [open, setOpen] = useState(false)
   const handleClose = () => {
     setOpen(false)
@@ -39,6 +37,9 @@ function DuplicationsPage(prop) {
     setOpen(!open)
   };
 
+  const jwtToken = localStorage.getItem("jwtToken")
+
+  //TODO 這邊寫死的記得要改唷!!!! >////<
 
   useEffect(() => {
     Axios.get(`http://localhost:9100/pvs-api/project/1/${projectId}`,
@@ -46,7 +47,7 @@ function DuplicationsPage(prop) {
       .then(response => {
         setCurrentProject(response.data)
       })
-      .catch((e) => {
+      .catch(e => {
         alert(e.response.status)
         console.error(e)
       })
@@ -58,11 +59,11 @@ function DuplicationsPage(prop) {
       let repositoryDTO = currentProject.repositoryDTOList.find(x => x.type == "sonar")
       let sonarToken = repositoryDTO.token
       let sonarComponent = repositoryDTO.url.split("id=")[1]
-      setDuplicationUrl(`http://localhost:9000/component_measures?id=${sonarComponent}&metric=Duplications&view=list`)
-      Axios.get(`http://localhost:9100/pvs-api/sonar/${sonarComponent}/duplication?token=${sonarToken}`,
+      setBugUrl(`http://localhost:9000/project/issues?id=${sonarComponent}&resolved=false&types=BUG`)
+      Axios.get(`http://localhost:9100/pvs-api/sonar/${sonarComponent}/bug?token=${sonarToken}`,
         { headers: { "Authorization": `${jwtToken}` } })
         .then((response) => {
-          setDuplicationList(response.data)
+          setBugList(response.data)
         })
         .catch((e) => {
           alert(e.response.status)
@@ -71,18 +72,19 @@ function DuplicationsPage(prop) {
     }
   }, [currentProject])
 
-  useEffect(() => {
-    let chartDataset = { labels: [], data: { duplication: [] } }
 
-    duplicationList.forEach(duplication => {
-      chartDataset.labels.push(moment(duplication.date).format("YYYY-MM-DD HH:mm:ss"))
-      chartDataset.data.duplication.push(duplication.value)
+  useEffect(() => {
+    let chartDataset = { labels: [], data: { bug: [] } }
+
+    bugList.forEach(bug => {
+      chartDataset.labels.push(moment(bug.date).format("YYYY-MM-DD HH:mm:ss"))
+      chartDataset.data.bug.push(bug.value)
     })
 
-    setDataForDuplicationChart(chartDataset)
+    setDataForBugChart(chartDataset)
     handleClose()
 
-  }, [duplicationList, prop.startMonth, prop.endMonth])
+  }, [bugList, prop.startMonth, prop.endMonth])
 
   return (
     <div style={{ marginLeft: "10px" }}>
@@ -98,14 +100,13 @@ function DuplicationsPage(prop) {
           <h2 id="number-of-sonar">{currentProject ? currentProject.projectName : ""}</h2>
         </p>
       </div>
-
       <div className={classes.root}>
         <div style={{ width: "67%" }}>
           <div>
-            <h1>Duplications</h1>
-            <h2><a href={duplicationUrl} target="blank">{dataForDuplicationChart.data.duplication[dataForDuplicationChart.data.duplication.length - 1]}%</a></h2>
+            <h1>Bugs</h1>
+            <h2><a href={bugUrl} target="blank">{dataForBugChart.data.bug[dataForBugChart.data.bug.length - 1]}</a></h2>
             <div>
-              <DrawingBoard data={dataForDuplicationChart} maxBoardY={100} id="duplications-chart" />
+              <DrawingBoard data={dataForBugChart} maxBoardY={Math.max(...dataForBugChart.data.bug) + 5} id="bugs-chart" />
             </div>
           </div>
         </div>
@@ -121,4 +122,4 @@ const mapStateToProps = (state) => {
   }
 }
 
-export default connect(mapStateToProps)(DuplicationsPage)
+export default connect(mapStateToProps)(BugsPage)
