+++ frontend/src/App/component/trello/TrelloCard.jsx
@@ -1,0 +1,16 @@
+import { Card } from "@material-ui/core";
+import React from "react";
+
+export default function TrelloCard(prop) {
+    return (
+        <div>
+            <div className="card">
+                <blockquote className="trello-card">
+                    <a href={prop.shortUrl}>
+                        {prop.shortUrl}
+                    </a>
+                </blockquote>
+            </div>
+        </div>
+    )
+}
