+++ frontend/src/routes/Routes.jsx
@@ -1,13 +1,13 @@
 import Login from '../App/component/Login'
 import SelectProject from '../App/component/SelectProject'
 import DashboardPage from '../App/component/DashboardPage'
-import CommitsPage from '../App/component/CommitsPage'
-import IssuesPage from '../App/component/IssuesPage'
-import CodeBasePage from '../App/component/CodeBasePage'
-import CodeCoveragePage from '../App/component/CodeCoveragePage'
-import BugsPage from '../App/component/BugsPage'
-import CodeSmellsPage from '../App/component/CodeSmellsPage'
-import DuplicationsPage from '../App/component/DuplicationsPage'
+import CommitsPage from '../App/component/github/CommitsPage'
+import IssuesPage from '../App/component/github/IssuesPage'
+import CodeBasePage from '../App/component/github/CodeBasePage'
+import CodeCoveragePage from '../App/component/sonar/CodeCoveragePage'
+import BugsPage from '../App/component/sonar/BugsPage'
+import CodeSmellsPage from '../App/component/sonar/CodeSmellsPage'
+import DuplicationsPage from '../App/component/sonar/DuplicationsPage'
 
 const routes = [
   { path: "/", redirect: true, to: "/select" },
